requirements.txt
Flask==3.0.2
flask-pymongo==2.3.0
python-dotenv==1.0.1
pymongo==4.6.1
requests==2.31.0 